requirements.txt
uvicorn
gpt3_tokenizer
pydantic
fastapi
openai
python-dotenv
firebase-admin
